requirements.txt
neo4j>=5.0.0
python-dotenv>=1.0.0
kg-gen>=0.1.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
